+++ assets/js/components/header.js
@@ -3,7 +3,8 @@ import { connect } from 'react-redux'
 import { NavLink } from 'react-router-dom'
 import { Container, Dropdown, Icon, Message, Menu } from 'semantic-ui-react'
 import { flashOperations, flashSelectors as Flash } from '../modules/flash'
-import { DefaultMenu } from './menus'
+import { simSelectors as Sim } from '../modules/sim'
+import { DefaultMenu, SimMenu } from './menus'
 
 class Header extends Component {
 
@@ -12,7 +13,7 @@ class Header extends Component {
   toggle = () => this.setState({ open: !this.state.open })
 
   render(){
-    const { clearFlash, icon, message, messageType, user } = this.props
+    const { clearFlash, icon, message, messageType, simulation, user } = this.props
     return([
       <Menu attached='bottom' inverted stackable key={0}>
         <Container>
@@ -24,7 +25,7 @@ class Header extends Component {
             onClick={() => this.toggle()}
             position='right'
           />
-          <DefaultMenu user={user} />
+          {simulation ? <SimMenu user={user} /> : <DefaultMenu user={user} />}
           <Menu.Menu position='right'>
             {!user && <Menu.Item as={NavLink} to='/login' icon='power' name='Login' />}
             {user &&
@@ -52,6 +53,7 @@ const mapStateToProps = store => ({
   icon: Flash.icon(store),
   message: Flash.message(store),
   messageType: Flash.messageType(store),
+  simulation: Sim.playerReady(store),
   user: store.user
 });
 const mapDispatchToProps = { clearFlash: flashOperations.clear }