+++ assets/js/components/menus/index.js
@@ -1,1 +1,2 @@
 export { default as DefaultMenu } from "./default_menu"
+export { default as SimMenu } from "./sim_menu"